fix_4847 Refactor jdbc source tasks
