+++ ohara-connector/src/test/scala/oharastream/ohara/connector/jdbc/source/TestJDBCSourceConnectorExactlyOnceTimestamp.scala
@@ -19,7 +19,7 @@ package oharastream.ohara.connector.jdbc.source
 import oharastream.ohara.kafka.connector.TaskSetting
 import scala.jdk.CollectionConverters._
 
-class TestJDBCSourceConnectorTimestamp(inputDataTime: Long)
+class TestJDBCSourceConnectorExactlyOnceTimestamp(inputDataTime: Long)
     extends BasicTestJDBCSourceConnectorExactlyOnce(inputDataTime: Long) {
   override protected[this] val jdbcSourceConnectorProps: JDBCSourceConnectorConfig = JDBCSourceConnectorConfig(
     TaskSetting.of(
