+++ ohara-connector/src/test/scala/oharastream/ohara/connector/jdbc/source/TestJDBCSourceTimestampTaskOffset.scala
@@ -34,13 +34,13 @@ import org.mockito.Mockito.when
 import org.scalatest.matchers.should.Matchers._
 import scala.jdk.CollectionConverters._
 
-class TestJDBCSourceTaskOffset extends OharaTest {
+class TestJDBCSourceTimestampTaskOffset extends OharaTest {
   private[this] val db                  = Database.local()
   private[this] val client              = DatabaseClient.builder.url(db.url()).user(db.user()).password(db.password()).build
   private[this] val tableName           = "TABLE1"
   private[this] val timestampColumnName = "COLUMN1"
 
-  private[this] val jdbcSourceTask: JDBCSourceTask           = new JDBCSourceTask()
+  private[this] val jdbcSourceTask: JDBCSourceTimestampTask  = new JDBCSourceTimestampTask()
   private[this] val taskContext: SourceTaskContext           = Mockito.mock(classOf[SourceTaskContext])
   private[this] val taskSetting: TaskSetting                 = Mockito.mock(classOf[TaskSetting])
   private[this] val offsetStorageReader: OffsetStorageReader = Mockito.mock(classOf[OffsetStorageReader])
